+++ MangoPay.SDK/Entities/POST/PayInPayPalWebPostDTO.cs
@@ -18,7 +18,9 @@ namespace MangoPay.SDK.Entities.POST
             Shipping shipping = null,
             string statementDescriptor = null,
             CultureCode? culture = null,
-            ShippingPreference? shippingPreference = null
+            ShippingPreference? shippingPreference = null,
+            string reference = null,
+            string cancelUrl = null
         )
         {
             AuthorId = authorId;
@@ -31,6 +33,8 @@ namespace MangoPay.SDK.Entities.POST
             StatementDescriptor = statementDescriptor;
             Culture = culture;
             ShippingPreference = shippingPreference;
+            Reference = reference;
+            CancelURL = cancelUrl;
         }
 
         /// <summary>Author identifier.</summary>
@@ -70,5 +74,9 @@ namespace MangoPay.SDK.Entities.POST
         /// Note that a particular bank may show more or less information.
         /// </summary>
         public string StatementDescriptor { get; set; }
+        
+        public string Reference { get; set; }
+        
+        public string CancelURL { get; set; }
     }
 }