+++ MangoPay.SDK/Entities/POST/RefundTransferPostDTO.cs
@@ -9,6 +9,13 @@ namespace MangoPay.SDK.Entities.POST
     /// <summary>Refund for Transfer POST entity.</summary>
     public class RefundTransferPostDTO : EntityPostBase
     {
+        public RefundTransferPostDTO(string authorId, Money fees, Money debitedFunds)
+        {
+            AuthorId = authorId;
+            Fees = fees;
+            DebitedFunds = debitedFunds;
+        }
+        
         public RefundTransferPostDTO(string authorId)
         {
             AuthorId = authorId;
@@ -16,5 +23,11 @@ namespace MangoPay.SDK.Entities.POST
 
         /// <summary>Author identifier.</summary>
         public string AuthorId { get; set; }
+        
+        /// <summary>Debited funds.</summary>
+        public Money DebitedFunds { get; set; }
+        
+        /// <summary>Fees.</summary>
+        public Money Fees { get; set; }
     }
 }